Migrated the assignment 3 code to a different file
